put credhub on port 80
we've recently moved our pipeline from our public ci cluster to the
vmware internal cluster. A side effect of that is that the datacenter
firewall blocks most ports other than a very few well known ones.

because the `deploy-it` task tries to hit credhub on port 8844, it gets
blocks and we are sad. so instead we expose credhub on port 80 (which
shouldn't be blocked).

we're not really sure who else is using this repo, so please create an
issue if this breaks your workflow.

Signed-off-by: Rui Yang <ruiya@vmware.com>
Co-authored-by: Bohan Chen <bochen@pivotal.io>